TOKO/requirements.txt.txt
Django==3.1.12
django-crispy-forms==1.9.0
Pillow==8.2.0
dj-database-url==0.3.0
dj-static==0.0.6
django-shopping-cart
